requirements.txt
Flask
openai
gunicorn
PyPDF2
docx2txt
pytesseract
Pillow
streamlit